requirements.txt
discord.py
openai==0.28
redis
python-dotenv
flask
schedule